+++ changelog.md
@@ -1,6 +1,12 @@
 # Changelogs
+## v1.0.2
+- 2.208 support (Geode v5)
+- Added 3 options: In Everywhere, In Gameplay and In Editor
+- Rewrote the entire mod
+
 ## v1.0.1
 Technical change, nothing interesting for the end user. 
 For developers: added the sprite to UILayer.
+
 ## v1.0.0
 Initial release